requirements.txt
opencv_python==4.9.0.80
protobuf==5.27.1
pyttsx3==2.90
SpeechRecognition==3.10.4
google-generativeai==0.6.0
PyAudio==0.2.12